Fixing bowls after order, spelling, change grammar
The bowls can now be altered when the user says 'no' to continuing to order.  There is one bug in that after every update, you then have to type 'no' again (working on a fix).  Spelling of one broth was changed to reflect the spell-checked menu Joselyn uploaded (across files), and grammar was updated to include adjective only order item for use after the bowls have been order (eg.  your response can be 'mild' to the question, 'how spicy did you want that?')