updated grammar dec 7
updated grammar from dec 7 that i was not able to pull request.
recursion for toppings ("beef and tofu"; "mushrooms and bean sprouts"),
a few more grammar additions ("miso ramen beef", "miso ramen with beef",
"miso ramen with beef and egg"), and addition of some prepositional
phrases "[noun] of [noun]". issues are that you can now order a "bottle
of ramen" or a "bowl of coke"